add option for brighter colors
colors 8 to 15 are brighter, the old method would darken colors 1 to 6 the new
method would be correct at brightening colors 9 to 14 however the old method
was kept and the ´--cols´ flag modified to accept the values of darken and
lighten.

for compatibility and expected behaviour the behaviour when the flag ´--cols16´
is passed without an argument has been kept as darkening colors 1 to 6.